requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.5.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.2
django-environ==0.12.0
djangorestframework==3.16.0
kombu==5.5.3
numpy==2.2.5
openpyxl==3.1.5
pandas==2.2.3
prompt_toolkit==3.0.51
python-dateutil==2.9.0.post0
pytz==2025.2
redis==6.0.0
six==1.17.0
sqlparse==0.5.3
style==1.1.0
tzdata==2025.2
update==0.0.1
vine==5.1.0
wcwidth==0.2.13
XlsxWriter==3.2.3
gunicorn
requests
